+++ Realm/RLMRealm.mm
@@ -265,6 +265,9 @@ void RLMRealmTranslateException(NSError **error) {
                 break;
         }
     }
+    catch (AddressSpaceExhausted const &ex) {
+        RLMSetErrorOrThrow(RLMMakeError(RLMErrorAddressSpaceExhausted, ex), error);
+    }
     catch (std::system_error const& ex) {
         RLMSetErrorOrThrow(RLMMakeError(ex), error);
     }
@@ -484,6 +487,10 @@ void RLMRealmTranslateException(NSError **error) {
     }
     catch (File::AccessError const& ex) {
         RLMSetErrorOrThrow(RLMMakeError(RLMErrorFail, ex), outError);
+        return NO;
+    }
+    catch (AddressSpaceExhausted const &ex) {
+        RLMSetErrorOrThrow(RLMMakeError(RLMErrorAddressSpaceExhausted, ex), outError);
         return NO;
     }
     catch (std::exception const& ex) {